1. 칸의 수 기준 랭킹 순위 추가
2. 영역의 수 기준 랭킹 순위 추가
3. 동점일때 순위 로직 처리

+++ src/main/java/com/dnd/ground/domain/exerciseRecord/service/ExerciseRecordService.java
@@ -23,7 +23,7 @@ public interface ExerciseRecordService {
 
     ResponseEntity<?> recordEnd(EndRequestDto endRequestDto);
 
-    Integer findMatrixNumber(List<ExerciseRecord> exerciseRecord);
+    Long findMatrixNumber(List<ExerciseRecord> exerciseRecord);
 
-    Integer findAreaNumber(List<ExerciseRecord> exerciseRecord);
+    Long findAreaNumber(List<ExerciseRecord> exerciseRecord);
 }
